feat(Team): add coverage

+++ app/graphql/types/position_coverage_type.rb
@@ -5,7 +5,7 @@ module Types
     description 'Positional coverage by a Player'
 
     Cap::POSITIONS.each do |pos|
-      field pos.to_sym, Integer,
+      field pos.to_sym, Float,
             "Coverage of #{pos} position", null: true
     end
   end
